Fix for [DEV-12621] check to see if migrate permissions has been run before starting webserver. Fixed typo in file name and scripts.
